Extract editor main toolbar text actions
