+++ assets/js/editor-main-toolbar-session.js
@@ -1,3 +1,5 @@
+import { createEditorMainToolbarTextActions } from './editor-main-toolbar-text-actions.js';
+
 const noop = () => {};
 const fallbackTranslate = (key) => key;
 
@@ -46,6 +48,10 @@ export function createEditorMainToolbarSession(options = {}) {
       }
       return null;
     };
+  const textActions = createEditorMainToolbarTextActions({
+    getEditorTextarea,
+    createInputEvent
+  });
 
   const editorToolbarEl = options.editorToolbarEl || getElementById('editorToolbar');
   const cardButton = options.cardButton || getElementById('btnInsertCard');
@@ -55,8 +61,6 @@ export function createEditorMainToolbarSession(options = {}) {
   const cardEmptyEl = options.cardEmptyEl || getElementById('cardPickerEmpty');
 
   let cardEntries = Array.isArray(options.cardEntries) ? options.cardEntries : [];
-  let lastSelectionRange = { start: 0, end: 0 };
-  let suppressSelectionTracking = false;
   let formattingButtons = [];
   let cardPopoverOpen = false;
   let cardPopoverClosing = false;
@@ -84,12 +88,6 @@ export function createEditorMainToolbarSession(options = {}) {
       return Array.isArray(entries) ? entries : [];
     }
     return Array.isArray(cardEntries) ? cardEntries : [];
-  };
-
-  const dispatchInputEvent = (textarea) => {
-    if (!textarea || typeof textarea.dispatchEvent !== 'function') return;
-    const event = createInputEvent();
-    if (event) textarea.dispatchEvent(event);
   };
 
   function applyButtonTooltipState(button, disabled) {
@@ -143,31 +141,10 @@ export function createEditorMainToolbarSession(options = {}) {
     applyButtonTooltipState(button, !!button.disabled);
   }
 
-  const isCaretOnEmptyLine = (textarea, selection) => {
-    if (!textarea || !selection) return false;
-    const { start, end } = selection;
-    if (end !== start) return false;
-    const value = textarea.value || '';
-    const lineStart = value.lastIndexOf('\n', start - 1) + 1;
-    let lineEnd = value.indexOf('\n', start);
-    if (lineEnd === -1) lineEnd = value.length;
-    const line = value.slice(lineStart, lineEnd);
-    return line.trim().length === 0;
-  };
-
-  const getNormalizedSelection = () => {
-    const textarea = getEditorTextarea();
-    if (!textarea) return { start: 0, end: 0 };
-    let start = textarea.selectionStart ?? 0;
-    let end = textarea.selectionEnd ?? start;
-    if (end < start) { const tmp = start; start = end; end = tmp; }
-    return { start, end };
-  };
-
   const updateFormattingToolbarState = () => {
     const textarea = getEditorTextarea();
-    const selection = lastSelectionRange || { start: 0, end: 0 };
-    const caretOnEmptyLine = isCaretOnEmptyLine(textarea, selection);
+    const selection = textActions.getLastSelection();
+    const caretOnEmptyLine = textActions.isCaretOnEmptyLine(textarea, selection);
     const hasSelection = selection.end > selection.start;
     formattingButtons.forEach(btn => {
       if (!btn || !btn.el) return;
@@ -192,202 +169,15 @@ export function createEditorMainToolbarSession(options = {}) {
   };
 
   const recordSelection = () => {
-    if (suppressSelectionTracking) return;
-    const textarea = getEditorTextarea();
-    if (!textarea) return;
-    lastSelectionRange = getNormalizedSelection();
+    if (!textActions.recordSelection()) return false;
     updateFormattingToolbarState();
+    return true;
   };
 
-  const restoreSelection = () => {
-    const textarea = getEditorTextarea();
-    if (!textarea) return null;
-    suppressSelectionTracking = true;
-    try {
-      try { textarea.focus(); }
-      catch (_) {}
-      if (lastSelectionRange) {
-        const { start, end } = lastSelectionRange;
-        if (typeof start === 'number' && typeof end === 'number') {
-          try { textarea.setSelectionRange(start, end); }
-          catch (_) {}
-        }
-      }
-    } finally {
-      suppressSelectionTracking = false;
-    }
-    return textarea;
-  };
-
-  const applyInlineFormat = (prefix, suffix) => {
-    const textarea = restoreSelection();
-    if (!textarea) return;
-    const { start, end } = getNormalizedSelection();
-    if (end <= start) return;
-    const value = textarea.value || '';
-    const selected = value.slice(start, end);
-    const startTag = String(prefix ?? '');
-    const endTag = String(suffix ?? '');
-    let replacement;
-    if (
-      selected.startsWith(startTag)
-      && selected.endsWith(endTag)
-      && selected.length >= startTag.length + endTag.length
-    ) {
-      replacement = selected.slice(startTag.length, selected.length - endTag.length);
-    } else {
-      replacement = `${startTag}${selected}${endTag}`;
-    }
-    textarea.setRangeText(replacement, start, end, 'end');
-    const newEnd = start + replacement.length;
-    textarea.setSelectionRange(start, newEnd);
-    dispatchInputEvent(textarea);
-    recordSelection();
-  };
-
-  const toggleLinePrefix = (prefix) => {
-    const textarea = restoreSelection();
-    if (!textarea) return;
-    const normalizedPrefix = String(prefix ?? '');
-    const selection = getNormalizedSelection();
-    let { start, end } = selection;
-    const wasCollapsed = end <= start;
-    const wasCaretOnEmptyLine = wasCollapsed && isCaretOnEmptyLine(textarea, selection);
-    const value = textarea.value || '';
-    if (end <= start) {
-      if (!wasCaretOnEmptyLine) return;
-      const lineStart = value.lastIndexOf('\n', start - 1) + 1;
-      let lineEnd = value.indexOf('\n', start);
-      if (lineEnd === -1) lineEnd = value.length;
-      start = lineStart;
-      end = lineEnd;
-    }
-    if (end < start) return;
-    const blockStart = value.lastIndexOf('\n', start - 1) + 1;
-    let blockEnd = value.indexOf('\n', end);
-    if (blockEnd === -1) blockEnd = value.length;
-    const block = value.slice(blockStart, blockEnd);
-    const lines = block.split('\n');
-    const shouldRemove = lines.every(line => {
-      const indentMatch = line.match(/^\s*/);
-      const indent = indentMatch ? indentMatch[0] : '';
-      return line.slice(indent.length).startsWith(normalizedPrefix);
-    });
-    const updated = lines.map(line => {
-      const indentMatch = line.match(/^\s*/);
-      const indent = indentMatch ? indentMatch[0] : '';
-      const content = line.slice(indent.length);
-      if (shouldRemove) {
-        if (content.startsWith(normalizedPrefix)) {
-          return indent + content.slice(normalizedPrefix.length);
-        }
-        return line;
-      }
-      if (content.startsWith(normalizedPrefix)) return line;
-      if (!content) return indent + normalizedPrefix;
-      return indent + normalizedPrefix + content;
-    });
-    const replacement = updated.join('\n');
-    textarea.setSelectionRange(blockStart, blockEnd);
-    textarea.setRangeText(replacement, blockStart, blockEnd, 'end');
-    const newEnd = blockStart + replacement.length;
-    if (wasCaretOnEmptyLine && wasCollapsed && !shouldRemove) {
-      const firstLine = replacement.split('\n', 1)[0] || '';
-      const indentMatch = firstLine.match(/^\s*/);
-      const indentLength = indentMatch ? indentMatch[0].length : 0;
-      const caretOffset = indentLength + normalizedPrefix.length;
-      const caretPos = blockStart + caretOffset;
-      textarea.setSelectionRange(caretPos, caretPos);
-    } else {
-      textarea.setSelectionRange(blockStart, newEnd);
-    }
-    dispatchInputEvent(textarea);
-    recordSelection();
-  };
-
-  const applyCodeBlockFormat = () => {
-    const textarea = restoreSelection();
-    if (!textarea) return;
-    const selection = getNormalizedSelection();
-    let { start, end } = selection;
-    const value = textarea.value || '';
-    if (end <= start) {
-      if (!isCaretOnEmptyLine(textarea, selection)) return;
-      const lineStart = value.lastIndexOf('\n', start - 1) + 1;
-      let lineEnd = value.indexOf('\n', start);
-      if (lineEnd === -1) lineEnd = value.length;
-      const beforeChar = lineStart > 0 ? value.charAt(lineStart - 1) : '';
-      const afterChar = lineEnd < value.length ? value.charAt(lineEnd) : '';
-      const prefix = beforeChar && beforeChar !== '\n' ? '\n' : '';
-      const suffix = afterChar && afterChar !== '\n' ? '\n' : '';
-      const block = '```\n\n```';
-      textarea.setSelectionRange(lineStart, lineEnd);
-      textarea.setRangeText(`${prefix}${block}${suffix}`, lineStart, lineEnd, 'end');
-      const caretPos = lineStart + prefix.length + 4;
-      textarea.setSelectionRange(caretPos, caretPos);
-      dispatchInputEvent(textarea);
-      recordSelection();
-      return;
-    }
-    const selected = value.slice(start, end);
-    const before = value.slice(0, start);
-    const after = value.slice(end);
-    let block = `\`\`\`\n${selected}\n\`\`\``;
-    let prefixAdded = false;
-    let suffixAdded = false;
-    if (start > 0 && !before.endsWith('\n')) {
-      block = `\n${block}`;
-      prefixAdded = true;
-    }
-    if (after && !after.startsWith('\n')) {
-      block = `${block}\n`;
-      suffixAdded = true;
-    }
-    textarea.setRangeText(block, start, end, 'end');
-    const selectionStart = start + (prefixAdded ? 1 : 0);
-    const selectionEnd = start + block.length - (suffixAdded ? 1 : 0);
-    textarea.setSelectionRange(selectionStart, Math.max(selectionStart, selectionEnd));
-    dispatchInputEvent(textarea);
-    recordSelection();
-  };
-
-  const insertCardLink = (entry) => {
-    if (!entry || !entry.location) return;
-    const location = String(entry.location).trim();
-    if (!location) return;
-    const textarea = restoreSelection();
-    if (!textarea) return;
-    const value = textarea.value || '';
-    const { start, end } = getNormalizedSelection();
-    const safeStart = Math.max(0, Math.min(start, value.length));
-    const safeEnd = Math.max(0, Math.min(end, value.length));
-    const hasSelection = safeEnd > safeStart;
-    const fallbackLabel = entry.key || entry.title || location;
-    let linkLabel = fallbackLabel;
-    if (hasSelection) {
-      const selected = value.slice(safeStart, safeEnd);
-      if (selected.trim()) linkLabel = selected;
-    }
-    const linkMarkdown = `[${linkLabel}](?id=${location})`;
-    let insertText = linkMarkdown;
-    let selectionStart = safeStart;
-    let selectionEnd = safeStart + linkMarkdown.length;
-    if (!hasSelection) {
-      const before = value.slice(0, safeStart);
-      const after = value.slice(safeStart);
-      const needsLeading = safeStart > 0 && !before.endsWith('\n');
-      const needsTrailing = after && !after.startsWith('\n');
-      const leading = needsLeading ? '\n' : '';
-      const trailing = needsTrailing ? '\n' : '';
-      insertText = `${leading}${linkMarkdown}${trailing}`;
-      selectionStart = safeStart + leading.length;
-      selectionEnd = selectionStart + linkMarkdown.length;
-    }
-    textarea.setSelectionRange(safeStart, safeEnd);
-    textarea.setRangeText(insertText, safeStart, safeEnd, 'end');
-    textarea.setSelectionRange(selectionStart, selectionEnd);
-    dispatchInputEvent(textarea);
-    recordSelection();
+  const runTextAction = (action) => {
+    const changed = typeof action === 'function' ? action() : false;
+    if (changed) updateFormattingToolbarState();
+    return changed;
   };
 
   const renderCardPickerList = (term = '') => {
@@ -421,7 +211,7 @@ export function createEditorMainToolbarSession(options = {}) {
       }
       btn.append(titleEl, metaEl);
       btn.addEventListener('click', () => {
-        insertCardLink(entry);
+        runTextAction(() => textActions.insertCardLink(entry));
         closeCardPopover();
       });
       frag.appendChild(btn);
@@ -475,7 +265,7 @@ export function createEditorMainToolbarSession(options = {}) {
     if (event.key === 'Escape') {
       event.preventDefault();
       closeCardPopover();
-      restoreSelection();
+      textActions.restoreSelection();
     }
   }
 
@@ -568,8 +358,8 @@ export function createEditorMainToolbarSession(options = {}) {
     updateFormattingToolbarState();
     const hasEntries = readCardEntries().length > 0;
     const textarea = getEditorTextarea();
-    const selection = lastSelectionRange || { start: 0, end: 0 };
-    const allowCardInsertion = hasEntries && isCaretOnEmptyLine(textarea, selection);
+    const selection = textActions.getLastSelection();
+    const allowCardInsertion = hasEntries && textActions.isCaretOnEmptyLine(textarea, selection);
     if ((!hasEntries || !allowCardInsertion) && cardPopoverOpen) {
       closeCardPopover();
       return;
@@ -583,17 +373,17 @@ export function createEditorMainToolbarSession(options = {}) {
   const selectionOrEmptyLineEnabled = (selection, textarea) => {
     if (!selection) return false;
     if (selection.end > selection.start) return true;
-    return isCaretOnEmptyLine(textarea, selection);
+    return textActions.isCaretOnEmptyLine(textarea, selection);
   };
 
   const formattingActions = [
-    { id: 'btnFmtBold', handler: () => applyInlineFormat('**', '**') },
-    { id: 'btnFmtItalic', handler: () => applyInlineFormat('*', '*') },
-    { id: 'btnFmtStrike', handler: () => applyInlineFormat('~~', '~~') },
-    { id: 'btnFmtHeading', handler: () => toggleLinePrefix('# '), isEnabled: selectionOrEmptyLineEnabled },
-    { id: 'btnFmtQuote', handler: () => toggleLinePrefix('> '), isEnabled: selectionOrEmptyLineEnabled },
-    { id: 'btnFmtCode', handler: () => applyInlineFormat('`', '`') },
-    { id: 'btnFmtCodeBlock', handler: () => applyCodeBlockFormat(), isEnabled: selectionOrEmptyLineEnabled }
+    { id: 'btnFmtBold', handler: () => textActions.applyInlineFormat('**', '**') },
+    { id: 'btnFmtItalic', handler: () => textActions.applyInlineFormat('*', '*') },
+    { id: 'btnFmtStrike', handler: () => textActions.applyInlineFormat('~~', '~~') },
+    { id: 'btnFmtHeading', handler: () => textActions.toggleLinePrefix('# '), isEnabled: selectionOrEmptyLineEnabled },
+    { id: 'btnFmtQuote', handler: () => textActions.toggleLinePrefix('> '), isEnabled: selectionOrEmptyLineEnabled },
+    { id: 'btnFmtCode', handler: () => textActions.applyInlineFormat('`', '`') },
+    { id: 'btnFmtCodeBlock', handler: () => textActions.applyCodeBlockFormat(), isEnabled: selectionOrEmptyLineEnabled }
   ];
 
   const bindFormattingButtons = () => {
@@ -603,7 +393,7 @@ export function createEditorMainToolbarSession(options = {}) {
       registerButtonTooltip(el, BUTTON_DISABLED_HINT_KEYS[action.id]);
       el.addEventListener('click', (event) => {
         event.preventDefault();
-        action.handler();
+        runTextAction(action.handler);
       });
       const requiresSelection = action.requiresSelection !== undefined ? action.requiresSelection : true;
       return { ...action, el, requiresSelection };
@@ -669,6 +459,6 @@ export function createEditorMainToolbarSession(options = {}) {
     setCardEntries,
     updateState: handleCardContextUpdate,
     recordSelection,
-    restoreSelection
+    restoreSelection: textActions.restoreSelection
   };
 }
