+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.88",
-  "tag": "v3.4.88",
+  "version": "3.4.89",
+  "tag": "v3.4.89",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.87 <3.4.88"
+      ">=3.4.88 <3.4.89"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.87 first."
+    "message": "Update to v3.4.88 first."
   }
 }
