+++ scripts/test-system-release-package.sh
@@ -323,6 +323,11 @@ if ! grep -qx "press-system-${version}/assets/js/editor-main-toolbar-session.js"
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/editor-main-toolbar-text-actions.js" "${entries_file}"; then
+  echo "expected package to include editor main toolbar text action code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-main-image-session.js" "${entries_file}"; then
   echo "expected package to include editor main image session boundary code" >&2
   exit 1
